added updates for NET4LLM new version 0.3.4
