+++ Demo/Demo.csproj
@@ -1,4 +1,5 @@
 <Project Sdk="Microsoft.NET.Sdk">
+  <!-- Demo: Support/, Samples/&lt;area&gt;/; MuPDF.NET4LLM fixtures in Samples/Llm/Program.Llm.*.Fixtures.cs. -->
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
@@ -7,8 +8,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MuPDF.NET\MuPDF.NET.csproj" />
     <ProjectReference Include="..\MuPDF.NET4LLM\MuPDF.NET4LLM.csproj" />
+    <ProjectReference Include="..\MuPDF.NET\MuPDF.NET.csproj" />
   </ItemGroup>
 
 </Project>
