+++ Demo/GlobalUsings.cs
@@ -1,0 +1,19 @@
+global using System;
+global using System.Collections.Generic;
+global using System.IO;
+global using System.Linq;
+global using System.Text;
+global using System.Threading;
+global using mupdf;
+global using MuPDF.NET;
+global using MuPDF.NET4LLM;
+global using MuPDF.NET4LLM.Helpers;
+global using MuPDF.NET4LLM.Llama;
+global using SkiaSharp;
+global using Box = MuPDF.NET.Box;
+global using Encoding = System.Text.Encoding;
+global using File = System.IO.File;
+global using Font = MuPDF.NET.Font;
+global using Morph = MuPDF.NET.Morph;
+global using TextWriter = MuPDF.NET.TextWriter;
+global using Utils = MuPDF.NET.Utils;
