+++ MuPDF.NET/MuPDF.NET.csproj
@@ -60,4 +60,13 @@
     <PackageReference Include="SkiaSharp" Version="3.119.1" />
   </ItemGroup>
 
+  <!-- MuPDF.NET.nuspec bundles MuPDF.NET4LLM.dll; ensure sibling project is built before Pack. -->
+  <!-- GeneratePackageOnBuild=false avoids nested Pack when 4LLM rebuilds MuPDF.NET. -->
+  <Target Name="BuildMuPDFNET4LLMForPack" BeforeTargets="Pack">
+    <MSBuild
+      Projects="$(MSBuildThisFileDirectory)..\MuPDF.NET4LLM\MuPDF.NET4LLM.csproj"
+      Targets="Build"
+      Properties="Configuration=$(Configuration);Platform=$(Platform);GeneratePackageOnBuild=false" />
+  </Target>
+
 </Project>
