+++ MuPDF.NET4LLM/LICENSE.md
@@ -1,0 +1,60 @@
+# ARTIFEX COMMUNITY LICENSE
+
+**Version 1, 11th December 2024**
+
+This **Artifex Community License** ("License") governs the use, reproduction, and distribution of **MuPDF.NET4LLM** ("Software") for non-commercial use. By using, modifying, or distributing this Software, you accept the terms of this License.
+
+You may not use the software without an appropriate license, so if you cannot abide by all the terms of this license, you must [contact the Licensor](https://artifex.com/contact) to discuss obtaining a **Commercial Use License**.
+
+## 1. Definitions
+
+- **Commercial Use**: Any use of the Software for commercial purposes, including, but not limited to, use as part of, or in any workflow in support of, a product, service, or platform that directly or indirectly generates revenue.
+- **Non-Commercial Use**: Any use of the software that does not fall into the **Commercial Use** category, such as for personal, or educational purposes.
+- **Licensor**: Artifex Software, Inc.
+- **You**: Any individual or entity using the Software.
+
+## 2. Grant of License
+
+1. **Non-Commercial Use**:
+	- You are granted a royalty-free, perpetual, non-exclusive license to use, modify, and distribute the Software for Non-Commercial Use, subject to the terms of this License.
+2. **Commercial Use**:
+	- This license does not cover Commercial Use.
+	- If you intend to use the Software for Commercial Use you are required to obtain a Commercial Use License - [contact the Licensor](https://artifex.com/contact) to negotiate a Commercial Use License.
+
+## 3. Permitted Uses
+
+1. You may:
+	- Access and modify the Software’s source code.
+	- Use the Software for any lawful purpose, subject to the limitations outlined in Section 2.
+	- Combine the Software with other works, provided that any distribution complies with this License.
+	- Distribute unmodified source code. 
+
+## 4. Restrictions
+
+1. You may not:
+	- Use the Software for any Commercial Use. If Commercial Use is required [contact the Licensor](https://artifex.com/contact) to negotiate a Commercial Use License.
+	- Remove or obscure any copyright, trademark, or attribution notices included in the Software.
+	- Distribute modified versions of the source code.
+
+2. Derivative works must include prominent attribution to the original Software and a copy of this License.
+
+## 5. Attribution
+
+All distributions of the Software, whether modified or unmodified, must include:
+- A prominent notice stating: *"This software is based on MuPDF.NET, developed by Artifex Software, Inc. and licensed under the Artifex Community License."*
+- A copy of this License.
+
+## 6. Disclaimer of Warranty
+
+THE SOFTWARE IS PROVIDED "AS IS," WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES, OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT, OR OTHERWISE, ARISING FROM, OUT OF, OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+## 7. Limitation of Liability
+
+IN NO EVENT SHALL THE LICENSOR BE LIABLE FOR ANY INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THE SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGES.
+
+## 8. Termination
+
+1. This License is effective until terminated.
+2. Your rights under this License will terminate automatically if you breach any of its terms. Upon termination, you must cease all use of the Software and destroy all copies in your possession.
+
+
