+++ MuPDF.NET4LLM/MuPDF.NET4LLM.nuspec
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="utf-8"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2013/05/nuspec.xsd">
+  <metadata>
+    <id>MuPDF.NET4LLM</id>
+    <version>$version$</version>
+    <authors>Artifex Software Inc.</authors>
+    <requireLicenseAcceptance>true</requireLicenseAcceptance>
+    <license type="file">LICENSE.md</license>
+    <icon>logo.png</icon>
+    <readme>Description.md</readme>
+    <projectUrl>https://github.com/ArtifexSoftware/MuPDF.NET</projectUrl>
+    <description>LLM/RAG helpers for MuPDF.NET: PDF-to-Markdown conversion, layout parsing, document structure analysis. Designed for use with RAG pipelines and integration with LLMs.</description>
+    <releaseNotes>LLM/RAG helpers for PDF processing: Markdown conversion, layout analysis, multi-column detection.</releaseNotes>
+    <tags>MuPDF PDF LLM RAG Markdown LlamaIndex document-processing PDF-to-text</tags>
+    <repository type="git" url="https://github.com/ArtifexSoftware/MuPDF.NET" />
+    <dependencies>
+      <group targetFramework=".NETFramework4.6.1">
+        <dependency id="MuPDF.NET" version="[3.2.13,)" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework=".NETFramework4.7.2">
+        <dependency id="MuPDF.NET" version="[3.2.13,)" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework=".NETFramework4.8">
+        <dependency id="MuPDF.NET" version="[3.2.13,)" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework="net5.0">
+        <dependency id="MuPDF.NET" version="[3.2.13,)" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework="net6.0">
+        <dependency id="MuPDF.NET" version="[3.2.13,)" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework="net7.0">
+        <dependency id="MuPDF.NET" version="[3.2.13,)" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework="net8.0">
+        <dependency id="MuPDF.NET" version="[3.2.13,)" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework=".NETStandard2.0">
+        <dependency id="MuPDF.NET" version="[3.2.13,)" exclude="Build,Analyzers" />
+      </group>
+    </dependencies>
+  </metadata>
+  <files>
+    <!-- $Configuration$ from NuspecProperties (Debug or Release) -->
+    <file src="bin\$Configuration$\netstandard2.0\MuPDF.NET4LLM.dll" target="lib\netstandard2.0\" />
+    <file src="bin\$Configuration$\net461\MuPDF.NET4LLM.dll" target="lib\net461\" />
+    <file src="bin\$Configuration$\net472\MuPDF.NET4LLM.dll" target="lib\net472\" />
+    <file src="bin\$Configuration$\net48\MuPDF.NET4LLM.dll" target="lib\net48\" />
+    <file src="bin\$Configuration$\net5.0\MuPDF.NET4LLM.dll" target="lib\net5.0\" />
+    <file src="bin\$Configuration$\net6.0\MuPDF.NET4LLM.dll" target="lib\net6.0\" />
+    <file src="bin\$Configuration$\net7.0\MuPDF.NET4LLM.dll" target="lib\net7.0\" />
+    <file src="bin\$Configuration$\net8.0\MuPDF.NET4LLM.dll" target="lib\net8.0\" />
+    <!-- Paths relative to this .nuspec (same folder as .csproj) -->
+    <file src="LICENSE.md" target="" />
+    <file src="Description.md" target="" />
+    <file src="logo.png" target="" />
+  </files>
+</package>
