+++ MuPDF.NET4LLM/VersionInfo.cs
@@ -7,6 +7,6 @@ namespace MuPDF.NET4LLM
     public static class VersionInfo
     {
         public static readonly (int Major, int Minor, int Patch) MinimumMuPDFVersion = (1, 27, 0);
-        public const string Version = "0.2.9";
+        public const string Version = "0.3.4-rc.1";
     }
 }
